11-Inner_Working_Function_Tools/Task11_Inner_Working_Function_Tools/pyproject.toml
[project]
name = "11-inner-working-function-tools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.1.1",
]
